Make sure stale, timed out logins don't cause crashes on later logins
Since the websocket connection outlived the login flow, it would still
receive messages from future attempts while not being in the right state
to handle them, causing a segfault and a crash. This ensures we don't
keep the old connections around and there's no way for them to cause
trouble.